+++ apps/backend-ops/src/emulator.ts
@@ -23,7 +23,7 @@ import { join } from 'path';
 import { backupBucket as prodBackupBucket, firebase as prodFirebase } from 'env/env.blockframes';
 import admin from 'firebase-admin'
 import { backupBucket, firebase } from '@env'
-import { migrateBeta } from './migrations';
+import { migrate } from './migrations';
 import { generateWatermarks, syncUsers } from './users';
 import { cleanDeprecatedData } from './db-cleaning';
 import { cleanStorage } from './storage-cleaning';
@@ -110,7 +110,7 @@ export async function anonDbProcess() {
   console.log(o.map((snap) => snap.id));
 
   console.info('Preparing database & storage by running migrations...');
-  await migrateBeta({ withBackup: false, db, storage }); // run the migration, do not trigger a backup before, since we already have it!
+  await migrate({ withBackup: false, db, storage }); // run the migration, do not trigger a backup before, since we already have it!
   console.info('Migrations complete!');
 
   console.log('Running anonymization...');
